Update tracing packages
