+++ platforms/intellij/skate/src/main/kotlin/foundry/intellij/skate/tracing/SkateTraceReporter.kt
@@ -21,14 +21,14 @@ import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.extensions.PluginId
 import com.intellij.openapi.project.Project
 import foundry.intellij.skate.util.tracingEndpoint
-import com.slack.sgp.tracing.KeyValue
-import com.slack.sgp.tracing.ListOfSpans
-import com.slack.sgp.tracing.model.buildSpan
-import com.slack.sgp.tracing.model.makeId
-import com.slack.sgp.tracing.model.newTagBuilder
-import com.slack.sgp.tracing.reporter.SimpleTraceReporter
-import com.slack.sgp.tracing.reporter.TraceReporter
-import com.slack.sgp.tracing.reporter.TraceReporter.NoOpTraceReporter
+import foundry.tracing.KeyValue
+import foundry.tracing.ListOfSpans
+import foundry.tracing.model.buildSpan
+import foundry.tracing.model.makeId
+import foundry.tracing.model.newTagBuilder
+import foundry.tracing.reporter.SimpleTraceReporter
+import foundry.tracing.reporter.TraceReporter
+import foundry.tracing.reporter.TraceReporter.NoOpTraceReporter
 import java.time.Duration
 import java.time.Instant
 import kotlin.time.DurationUnit
