+++ tools/tracing/gradle.properties
@@ -1,3 +1,3 @@
-POM_ARTIFACT_ID=sgp-tracing
-POM_NAME=SGP Tracing
-POM_DESCRIPTION=SGP Tracing
+POM_ARTIFACT_ID=tracing
+POM_NAME=Foundry Tracing
+POM_DESCRIPTION=Foundry Tracing
